Small layout changes, code deduplication
- Add extensions for duplicated code in dialogs with a text input.
- Remove "Show password" checkbox, instead use the builtin password toggle icon.
- Change IME action in import password dialog.
- Remove deprecated LayoutInflater.fromContext.
- Shorten encrypted export summary (the encryption algorithm doesn't matter much to the average user).
