+++ app/src/main/kotlin/com/maltaisn/notes/ui/labels/LabelEditDialog.kt
@@ -19,7 +19,6 @@ package com.maltaisn.notes.ui.labels
 import android.app.Dialog
 import android.content.DialogInterface
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.WindowManager
 import androidx.core.widget.doAfterTextChanged
 import androidx.fragment.app.DialogFragment
@@ -28,7 +27,9 @@ import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.maltaisn.notes.App
 import com.maltaisn.notes.R
 import com.maltaisn.notes.databinding.DialogLabelEditBinding
+import com.maltaisn.notes.hideCursorInAllViews
 import com.maltaisn.notes.model.entity.Label
+import com.maltaisn.notes.setTitleIfEnoughSpace
 import com.maltaisn.notes.ui.SharedViewModel
 import com.maltaisn.notes.ui.navGraphViewModel
 import com.maltaisn.notes.ui.observeEvent
@@ -56,7 +57,7 @@ class LabelEditDialog : DialogFragment() {
 
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
         val context = requireContext()
-        val binding = DialogLabelEditBinding.inflate(LayoutInflater.from(context), null, false)
+        val binding = DialogLabelEditBinding.inflate(layoutInflater, null, false)
 
         // Using `this` as lifecycle owner, cannot show dialog twice with same instance to avoid double observation.
         debugCheck(!viewModel.setLabelEvent.hasObservers()) { "Dialog was shown twice with same instance." }
@@ -71,18 +72,11 @@ class LabelEditDialog : DialogFragment() {
                 viewModel.addLabel()
             }
             .setNegativeButton(R.string.action_cancel, null)
-
-        // Only show dialog title if screen size is under a certain dimension.
-        // Otherwise it becomes much harder to type text, see #53.
-        val dimen = context.resources.getDimension(R.dimen.label_edit_dialog_title_min_height) /
-                context.resources.displayMetrics.density
-        if (context.resources.configuration.screenHeightDp >= dimen) {
-            builder.setTitle(if (args.labelId == Label.NO_ID) {
+            .setTitleIfEnoughSpace(if (args.labelId == Label.NO_ID) {
                 R.string.label_create
             } else {
                 R.string.label_edit
             })
-        }
 
         val dialog = builder.create()
 
@@ -103,13 +97,7 @@ class LabelEditDialog : DialogFragment() {
             viewModel.onHiddenChanged(isChecked)
         }
 
-        // Cursor must be hidden when dialog is dimissed to prevent memory leak
-        // See [https://stackoverflow.com/questions/36842805/dialogfragment-leaking-memory]
         nameInput.isCursorVisible = true
-        dialog.setOnDismissListener {
-            nameInput.isCursorVisible = false
-        }
-
         nameInput.doAfterTextChanged {
             viewModel.onNameChanged(it?.toString() ?: "")
         }
@@ -125,5 +113,10 @@ class LabelEditDialog : DialogFragment() {
         viewModel.start(args.labelId)
 
         return dialog
+    }
+
+    override fun onDismiss(dialog: DialogInterface) {
+        super.onDismiss(dialog)
+        hideCursorInAllViews()
     }
 }
