+++ app/src/main/res/layout/dialog_import_password.xml
@@ -1,3 +1,19 @@
+<!--
+  Copyright 2023 Nicolas Maltais
+
+  Licensed under the Apache License, Version 2.0 (the "License");
+  you may not use this file except in compliance with the License.
+  You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+  -->
+
 <ScrollView
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
@@ -9,49 +25,31 @@
     app:backgroundInsetTop="30dp"
     >
 
-    <LinearLayout
+    <com.google.android.material.textfield.TextInputLayout
+        android:id="@+id/password_input_layout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="vertical"
         android:paddingTop="8dp"
         android:paddingStart="16dp"
         android:paddingEnd="16dp"
+        app:passwordToggleEnabled="true"
         >
 
-        <!-- The dialog is unusable in landscape mode
-             with keyboard shown, but there's not much to do about it. -->
-
-        <com.google.android.material.textfield.TextInputLayout
-            android:id="@+id/password_input_layout"
+        <com.google.android.material.textfield.TextInputEditText
+            android:id="@+id/password_input"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            >
-
-            <com.google.android.material.textfield.TextInputEditText
-                android:id="@+id/password_input"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:paddingTop="16dp"
-                android:paddingBottom="16dp"
-                android:imeOptions="actionNext"
-                android:singleLine="true"
-                android:maxLines="1"
-                android:nextFocusDown="@android:id/button1"
-                android:textDirection="anyRtl"
-                android:hint="@string/encrypted_export_password"
-                android:inputType="textPassword"
-                tools:text="Password"
-                />
-        </com.google.android.material.textfield.TextInputLayout>
-
-        <CheckBox
-            android:id="@+id/show_password_chk"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
-            android:text="@string/encrypted_export_show_password"
-            android:textAlignment="viewStart"
+            android:paddingTop="16dp"
+            android:paddingBottom="16dp"
+            android:imeOptions="actionDone"
+            android:singleLine="true"
+            android:maxLines="1"
+            android:nextFocusDown="@android:id/button1"
+            android:textDirection="anyRtl"
+            android:hint="@string/encrypted_export_password"
+            android:inputType="textPassword"
+            tools:text="Password"
             />
+    </com.google.android.material.textfield.TextInputLayout>
 
-    </LinearLayout>
 </ScrollView>
