+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <!--
-  Copyright 2022 Nicolas Maltais
+  Copyright 2023 Nicolas Maltais
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -205,7 +205,8 @@
     <string name="pref_data_export">Export data</string>
     <string name="pref_data_export_summary">Export notes data as JSON</string>
     <string name="pref_data_encrypted_export">Encrypted export</string>
-    <string name="pref_data_encrypted_export_summary">Encrypts the exported notes with AES-256GCM. Access is not possible without the correct password.</string>
+    <string name="pref_data_encrypted_export_summary">Encrypts the exported notes. Access is not possible without
+        the correct password.</string>
     <string name="pref_data_auto_export">Automatic data export</string>
     <string name="pref_data_auto_export_summary">Automatically export the app\'s data to a file, once a day.</string>
     <string name="pref_data_auto_export_date">(Last exported on %s)</string>
@@ -222,16 +223,17 @@
     <string name="pref_about_version">Version</string>
 
     <string name="pref_restart_dialog_title">Restart required!</string>
-    <string name="pref_restart_dialog_description">A restart is required for these changes to take effect. Do you want to restart the app now?</string>
+    <string name="pref_restart_dialog_description">A restart is required for these changes to take effect.
+        Do you want to restart the app now?</string>
 
     <!-- Import/export data messages -->
     <string name="export_success">Data exported successfully!</string>
     <string name="export_fail">An error occurred during data export.</string>
-    <string name="export_serialization_fail">An error occurred during data export. Try disabling encryption and try again.</string>
-    <string name="encrypted_export_dialog_title">Set Password</string>
+    <string name="export_serialization_fail">An error occurred during data export.
+        Try disabling encryption and try again.</string>
+    <string name="encrypted_export_dialog_title">Set password</string>
     <string name="encrypted_export_password">Password</string>
-    <string name="encrypted_export_repeat_password">Confirm Password</string>
-    <string name="encrypted_export_show_password">Show password</string>
+    <string name="encrypted_export_repeat_password">Confirm password</string>
     <string name="encrypted_export_password_matching_error">Passwords do not match</string>
     <string name="encrypted_export_key_generation_successful">Encryption enabled successfully!</string>
     <string name="encrypted_export_key_deletion_successful">Encryption disabled successfully!</string>
@@ -243,8 +245,10 @@
     <string name="import_bad_data">Could not import data, some fields have invalid values.
         This can happen if importing data from a later version, try updating the app.</string>
     <string name="encrypted_import_dialog_title">Enter Password</string>
-    <string name="encrypted_import_key_error">An error occurred while attempting to decrypt this file. Check if the password was correct.</string>
-    <string name="encrypted_import_encryption_unsupported">Could not import data, encryption is not supported on this version of Android.</string>
+    <string name="encrypted_import_key_error">An error occurred while attempting to decrypt this file.
+        Check if the password was correct.</string>
+    <string name="encrypted_import_encryption_unsupported">Could not import data, encryption is not supported
+        on this version of Android.</string>
     <string name="auto_export_message">To enable automatic export, a file must be chosen to
         save the data. The app\'s data will be exported once a day while the app is opened.
         Click OK to proceed.</string>
